pitch-parser/pitch-parser-for-git/requirements.txt
e2b-code-interpreter>=0.0.9
python-dotenv>=1.0.0
